% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasaurus_dozen.R
\docType{data}
\name{datasaurus_dozen}
\alias{datasaurus_dozen}
\title{Datasaurus Dozen Data}
\format{
A data frame with 1846 rows and 3 variables:
\itemize{
\item \code{dataset}: the dataset the values come from
\item \code{x}: the x-variable
\item \code{y}: the y-variable
}
}
\usage{
datasaurus_dozen
}
\description{
A dataset containing 12 datasets that are equal in mean, variance, and
Pearson's correlation but very different when visualized.
}
\references{
Davies R, Locke S, D'Agostino McGowan L (2022). \emph{datasauRus: Datasets from the
Datasaurus Dozen}. R package version 0.1.6, \url{https://CRAN.R-project.org/package=datasauRus}.

Matejka, J., & Fitzmaurice, G. (2017). Same Stats, Different Graphs:
Generating Datasets with Varied Appearance and Identical Statistics through
Simulated Annealing. CHI 2017 Conference proceedings: ACM SIGCHI Conference
on Human Factors in Computing Systems. Retrieved from https://www.autodesk.com/research/publications/same-stats-different-graphs
}
\keyword{datasets}
