% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_causal_quartet.R
\docType{data}
\name{variation_causal_quartet}
\alias{variation_causal_quartet}
\alias{variation_conceptual_causal_quartet}
\title{Gelman Variation Causal Quartet Data}
\format{
\code{variation_causal_quartet}: A dataframe with 88 rows and 4 variables:
\itemize{
\item \code{dataset}: The data generating mechanism
\item \code{x}: exposure
\item \code{z}: a pre-exposure factor
\item \code{y}: outcome
}

\code{variation_conceptual_causal_quartet}:  A dataframe with 88 rows and 4 variables:
\itemize{
\item \code{dataset}: The data generating mechanism
\item \code{z}: A pre-exposure factor
\item \code{causal_effect}: The true causal effect
}
}
\usage{
variation_causal_quartet

variation_conceptual_causal_quartet
}
\description{
This dataset contains 88 observations, each generated under a different
mechanism of variation of the treatment effect with respect to some
pre-exposure characteristic, \code{z}:
\itemize{
\item (1) Constant effect
\item (2) Low variation
\item (3) High variation
\item (4) Occasional large effects
}
}
\details{
The dataframe \code{variation_conceptual_causal_quartet} contains the latent
"true" causal effect.
}
\references{
Gelman, A., Hullman, J., & Kennedy, L. (2023). Causal quartets: Different ways to attain the same average treatment effect. arXiv preprint arXiv:2302.12878.

Hullman J (2023). \emph{causalQuartet: Create Causal Quartets for Interrogating Average Treatment Effects}. R package version 0.0.0.9000.
}
\keyword{datasets}
