% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc_quartet.R
\docType{data}
\name{auroc_quartet}
\alias{auroc_quartet}
\title{Area under the ROC Curve (AUROC) Quartet Data}
\format{
A dataframe with 400,000 rows and 4 variables:
\itemize{
\item \code{id}: The data id number within each method
\item \code{method}: The data generating mechanism ("beta" or "piecewise")
\item \code{score}: The numerical prediction score
\item \code{target}: The true binary outcome being predicted
}
}
\usage{
auroc_quartet
}
\description{
This dataset contains 4 datasets, each with equivalent
values for the AUROC, a popular model evaluation metric for
classification problems. Each curve features a substantially
different AUROC curve shape with implications for other evaluation
methods such as lift, precision, and recall.
}
\details{
AUROC curves are parameterized as one of two functional forms
defined by \code{method} (piecewise linear or the Beta distribution CDF).
}
\references{
Riederer E (2026). \emph{roctet}: . Python package version 0.1.0.
}
\keyword{datasets}
