% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_collider.R
\docType{data}
\name{causal_collider}
\alias{causal_collider}
\title{Collider Data}
\format{
A dataframe with 100 rows and 3 variables:
\itemize{
\item \code{x}: exposure
\item \code{y}: outcome
\item \code{z}: a known factor (collider)
}
}
\usage{
causal_collider
}
\description{
This dataset contains 100 observations, generated under the following mechanism:
X ~ N(0, 1)
Y ~ X + N(0,1)
Z ~ 0.45X + 0.77Y + N(0, 1)
}
\keyword{datasets}
