% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_confounding.R
\docType{data}
\name{causal_confounding}
\alias{causal_confounding}
\title{Confounder Data}
\format{
A dataframe with 100 rows and 3:
\itemize{
\item \code{z}: a known factor (confounder)
\item \code{x}: exposure
\item \code{y}: outcome
}
}
\usage{
causal_confounding
}
\description{
This dataset contains 100 observations, generated under the following mechanism:
Z ~ N(0, 1)
X ~ Z + N(0,1)
Y ~ 0.5X + Z + N(0, 1)
}
\keyword{datasets}
