% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_m_bias.R
\docType{data}
\name{causal_m_bias}
\alias{causal_m_bias}
\title{M-Bias Data}
\format{
A dataframe with 100 rows and 5 variables:
\itemize{
\item \code{u1}: an unknown factor
\item \code{u2}: an unknown factor
\item \code{z}: a known factor
\item \code{x}: exposure
\item \code{y}: outcome
}
}
\usage{
causal_m_bias
}
\description{
This dataset contains 100 observations, generated under the following mechanism:
U1 ~ N(0, 1)
U2 ~ N(0, 1)
Z ~ 8 U1 + U2 + N(0, 1)
X ~ U1 + N(0,1)
Y ~ X + U2 + N(0, 1)
}
\keyword{datasets}
