% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_mediator.R
\docType{data}
\name{causal_mediator}
\alias{causal_mediator}
\title{Mediator Data}
\format{
A dataframe with 100 rows and 3 variables:
\itemize{
\item \code{exposure}: exposure
\item \code{covariate}: a known factor (mediator)
\item \code{outcome}: outcome
}
}
\usage{
causal_mediator
}
\description{
This dataset contains 100 observations, generated under the following mechanism:
X ~ N(0, 1) (exposure)
Z ~ X + N(0,1) (measured factor: mediator)
Y ~ Z + N(0, 1) (outcome)
}
\references{
D'Agostino McGowan L, Barrett M (2023). Causal inference is not a statistical problem. Preprint arXiv:2304.02683v1.
}
\keyword{datasets}
