% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_quartet.R
\docType{data}
\name{causal_quartet}
\alias{causal_quartet}
\title{Causal Quartet Data}
\format{
A dataframe with 400 rows and 6 variables:
\itemize{
\item \code{dataset}: The data generating mechanism
\item \code{x}: exposure
\item \code{y}: outcome
\item \code{z}: a known factor
\item \code{u1}: an unknown factor
\item \code{u2}: an unknown factor
}
}
\usage{
causal_quartet
}
\description{
This dataset contains 400 observations, each generated under a different
data generating mechanism:
\itemize{
\item (1) A collider
\item (2) A confounder
\item (3) A mediator
\item (4) M-bias
}
}
\keyword{datasets}
