% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_causal_quartet.R
\docType{data}
\name{variation_causal_quartet}
\alias{variation_causal_quartet}
\title{Gelman Variation Causal Quartet Data}
\format{
A dataframe with 88 rows and 5 variables:
\itemize{
\item \code{dataset}: The data generating mechanism
\item \code{exposure}: exposure
\item \code{covariate}: a pre-exposure factor
\item \code{outcome}: outcome
\item \code{.causal_effect}: Latent true causal effect
}
}
\usage{
variation_causal_quartet
}
\description{
This dataset contains 88 observations, each generated under a different
mechanism of variation of the treatment effect with respect to some
pre-exposure characteristic, \code{z}:
\itemize{
\item (1) Constant effect
\item (2) Low variation
\item (3) High variation
\item (4) Occasional large effects
}
}
\references{
Gelman, A., Hullman, J., & Kennedy, L. (2023). Causal quartets: Different ways to attain the same average treatment effect. arXiv preprint arXiv:2302.12878.

Hullman J (2023). \emph{causalQuartet: Create Causal Quartets for Interrogating Average Treatment Effects}. R package version 0.0.0.9000.
}
\keyword{datasets}
